Shitty optimization
Significantly reduces the amount of work GT does when checking whether new
recipes are duplicates, and probably other places as well. Saves noticable
time during startup. However, the implementation here is kinda poor, and
there are still savings to be made. So this commit is more about sharing a
prototype with others than seriously intended to be merged as-is.